Move potential functions in .\minplascal\transport
